+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Bracelets (requires Curios). Functionally identical to rings, but they go in the bracelet slots.
+### Fixed
+- (Partial fix) Gear crossbows destroying gear arrows. Vanilla crossbows not fixed. [#270]
+
 ## [2.4.0] - 2020-11-29
 ### Added
 - Rings. Requires Curios to function correctly. The shank can only be crafted with metals (this can be overridden with data packs if desired).